refactor(components): split RepoResults into separate components
- Move tab bar and results container into separate components
- Update import paths to use new component structure- Refactor HeaderControls to adjust prop order
- Remove unused dev tools imports
- Update ProgressTracker to import Zoom component correctly
- Adjust PullRequestsTab to use correct import path
- Update RepoActions button styling and functionality
