+++ src/components/github/repo-analysis/index.ts
@@ -1,0 +1,4 @@
+export { default as RepoResults } from './RepoResults';
+export { default as RepoTabBar } from './RepoTabBar';
+export { default as RepoResultsContainer } from './RepoResultsContainer';
+export { default as TabPanel } from './TabPanel';
