--- src/dev-tools/TanStackTableDevTools.tsx
@@ -1,10 +1,0 @@
-import React from "react";
-import { isProduction } from "../utils";
-
-export const TanStackTableDevTools = isProduction
-	? (): null => null
-	: React.lazy(() =>
-			import("@tanstack/react-table-devtools").then((result) => ({
-				default: result.ReactTableDevtools,
-			}))
-	  );
