--- src/dev-tools/index.ts
@@ -1,3 +1,0 @@
-export { ReactHookFormDevTools } from './ReactHookFormDevTools';
-export { TanStackRouterDevTools } from './TanStackRouterDevTools';
-export { TanStackTableDevTools } from './TanStackTableDevTools';
